Automatic display of productos from database
Display of thumbnails of products is no more an static process. It takes
the elements form the web service. Ajax is used to read data from
dataservice and to load de images.
Button home display is changed to fix, and media is divided in screen
with more the 800px and less than 800px.